Implement Block Nested Loop algorithm for cross/inner joins
The block nested loop algorithm is an extension to the nested loop
algorithm. Instead of processing the items of the right relation once for every
item of the left relation, it builds blocks of item of the left relation and
processes the blocks for each item on the right.

This saves IO operations and generally it is expected for the performance to
increase. Further benchmarks will show if that holds true. One drawback of the
new algorithm, is order of the left relation can't be preserved. Thus, pushing
down an order by is not possible in case of a block nested loop execution.

The following interfaces were created in the to keep the
CrossJoinNLBatchIterator in the :dex module.

- BlockSizecalculator
- RamAccounting
- RowAccounting